requirements[1].txt
python-telegram-bot==13.15
python-dotenv==1.0.1